+++ lib/commands/8ball.ts
@@ -10,17 +10,17 @@ export const command: Classes.Command = new Classes.Command({
 	exp: /^!8(ball)?( .+)?$/smi,
 	category: "Utility",
 	data: {
-		cache: new Classes.CacheBank("8ball", null, true, false)
+		cache: new Classes.CacheBank("8ball", undefined, true, false, "https://nekos.life/api/v2/8ball")
 	}, 
 	body: async function body(message: Message, vale?: Classes.Vale): Promise<void> {
 		let repl = Classes.failsafe.bind(message);
 
 		try {
-			let reply: string = this.data.cache.get() || await Classes.fetch("https://nekos.life/api/v2/8ball"),
+			let reply: string = this.data.cache.get() || await Classes.fetch(command.data.cache.source),
 				embed: RichEmbed = new RichEmbed(),
 				send = JSON.parse(reply);
 			
-			Classes.fetch("https://nekos.life/api/v2/8ball").then((reply: string) => command.data.cache.push(reply));
+			Classes.fetch(command.data.cache.source).then((reply: string) => command.data.cache.push(reply));
 					
 			embed.setImage(send.url)
 				.setURL(send.url)