+++ lib/commands/chat.ts
@@ -7,7 +7,7 @@ export const command: Classes.Command = new Classes.Command({
 	name: "chat",
 	desc: "Chat with me! :3",
 	usage: "chat",
-	exp: /^!chat( .*)?$/msi,
+	exp: /^!chat .+$/msi,
 	category: "Utility",
 	data: { },
 	body: async function body(message: Message, vale?: Classes.Vale): Promise<void> {
@@ -16,7 +16,7 @@ export const command: Classes.Command = new Classes.Command({
 		try {
 			let reply: string = await Classes.fetch("https://nekos.life/api/v2/chat?text=" + message.content.split(' ').slice(1).join(' '));
 
-			repl(JSON.parse(reply).response.replace(/<\d+>/g, ''));
+			repl(decodeURIComponent(JSON.parse(reply).response).replace(/<.?\d+>/g, ''));
 		} catch (err) {
 			repl("External API error, please try again later... https://nekos.life/api/v2/endpoints");
 			console.error(err);
@@ -26,7 +26,7 @@ export const command: Classes.Command = new Classes.Command({
 
 export async function init(vale: Classes.Vale): Promise<Classes.Command> {
 	command.usage = vale.opts.config.prefix + command.usage;
-	command.exp = new RegExp('^' + vale.opts.config.prefix + "chat( .*)?$", "msi");
+	command.exp = new RegExp('^' + vale.opts.config.prefix + "chat .+$", "msi");
 
 	return command;
 } //init